refactor(gyms): 🚚 Move Gym modals to the "Modals" folder
