+++ src/components/Modals/Gyms/ModalGym.tsx
@@ -2,11 +2,11 @@ import { images } from '@src/utils/utils';
 import React, { useContext, useEffect, useState } from 'react';
 import Modal from 'react-native-modal';
 import { View, Text, FlatList, StyleSheet, Image } from 'react-native';
-import { CustomButton } from '../CustomButton';
+import { CustomButton } from '../../CustomButton';
 import { ModalRewards } from './ModalRewards';
 import { GymsModalContext } from '@src/context/GymsModalContext';
 import { requestGymInfoById } from '@src/services/map.services';
-import { TGymInfo, TGymLoomieProtector, TReward } from '../../types/types';
+import { TGymInfo, TGymLoomieProtector, TReward } from '../../../types/types';
 import { getPosition } from '@src/services/geolocation.services';
 import { requestRewards } from '@src/services/map.services';
 import { useToastAlert } from '@src/hooks/useToastAlert';
