+++ src/pages/MapView.tsx
@@ -5,7 +5,7 @@ import { useAuth } from '../hooks/useAuth';
 import { useToastAlert } from '../hooks/useToastAlert';
 import { Map3D } from '@src/components/Map3D/Map3D';
 import FeatherIcon from 'react-native-vector-icons/Feather';
-import { ModalGym } from '@src/components/ModalDialogs/ModalGym';
+import { ModalGym } from '@src/components/Modals/Gyms/ModalGym';
 import { GymsModalProvider } from '@src/context/GymsModalContext';
 import { BabylonContext } from '@src/context/BabylonProvider';
